+++ packages/crypto/src/node/aes-ctr.ts
@@ -2,21 +2,13 @@ import { arrayify } from '@ethersproject/bytes';
 import { pbkdf2 } from '@ethersproject/pbkdf2';
 import crypto from 'crypto';
 
-import type { CryptoApi, Encoding, Keystore } from '../types';
+import type { CryptoApi, Keystore } from '../types';
 
+import { bufferFromString } from './bufferFromString';
 import { randomBytes } from './randomBytes';
+import { stringFromBuffer } from './stringFromBuffer';
 
 const ALGORITHM = 'aes-256-ctr';
-
-export const bufferFromString: CryptoApi['bufferFromString'] = (
-  string: string,
-  encoding: Encoding = 'base64'
-): Uint8Array => Buffer.from(string, encoding);
-
-export const stringFromBuffer: CryptoApi['stringFromBuffer'] = (
-  buffer: Uint8Array,
-  encoding: Encoding = 'base64'
-): string => Buffer.from(buffer).toString(encoding);
 
 /**
  * Generate a pbkdf2 key from a password and random salt
