+++ packages/crypto/src/node/stringFromBuffer.ts
@@ -1,0 +1,6 @@
+import type { CryptoApi, Encoding } from '../types';
+
+export const stringFromBuffer: CryptoApi['stringFromBuffer'] = (
+  buffer: Uint8Array,
+  encoding: Encoding = 'base64'
+): string => Buffer.from(buffer).toString(encoding);
